+++ src/app/highscore/highscore.component.ts
@@ -1,9 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { HighscoreEntry } from './../models/highscore-entry';
+import { HighscoreEntry, Level } from '../models';
 import { Router } from '@angular/router';
-import { HighscoreService } from '../services/highscore.service';
-import { LevelService } from '../services/level.service';
-import { Level } from '../models/level';
+import { HighscoreService, LevelService } from '../services';
 
 @Component({
   selector: 'app-highscore',
@@ -20,7 +18,7 @@ export class HighscoreComponent implements OnInit {
     this.entries = service.getLevel(this.index);
   }
 
-  ngOnInit() {}
+  ngOnInit() { }
 
   next() {
     this.index = this.index === this.levelService.getLevelCount() - 1 ? 0 : this.index + 1;
