+++ src/app/models/index.ts
@@ -1,0 +1,9 @@
+export * from './binaryHeap';
+export * from './closedList';
+export * from './coordinate';
+export * from './direction';
+export * from './highscore-entry';
+export * from './level';
+export * from './openList';
+export * from './savegame';
+export * from './tile';
