+++ src/app/services/highscore.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
-import { HighscoreEntry } from '../models/highscore-entry';
+import { HighscoreEntry } from '../models';
 
 @Injectable({
   providedIn: 'root'
 })
 export class HighscoreService {
-  constructor() {}
+  constructor() { }
 
   addEntry(index: number, entry: HighscoreEntry): void {
     const entries = this.getLevel(index);
