+++ src/app/services/index.ts
@@ -1,0 +1,4 @@
+export * from './highscore.service';
+export * from './level.service';
+export * from './pathFinder.service';
+export * from './mobileZoom.service';
